Move validating webhook config to a separate file
